Add `:legacy:notification` module and move notification light, settings and vibration there
