+++ legacy/core/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
     api(projects.core.android.common)
 
     api(projects.legacy.account)
+    api(projects.legacy.notification)
 
     implementation(projects.plugins.openpgpApiLib.openpgpApi)
