+++ legacy/core/src/test/java/com/fsck/k9/notification/BaseNotificationDataCreatorTest.kt
@@ -1,6 +1,9 @@
 package com.fsck.k9.notification
 
 import app.k9mail.legacy.account.Identity
+import app.k9mail.legacy.notification.NotificationLight
+import app.k9mail.legacy.notification.NotificationVibration
+import app.k9mail.legacy.notification.VibratePattern
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isInstanceOf
@@ -9,9 +12,6 @@ import assertk.assertions.isSameInstanceAs
 import com.fsck.k9.Account
 import com.fsck.k9.K9
 import com.fsck.k9.K9.LockScreenNotificationVisibility
-import com.fsck.k9.NotificationLight
-import com.fsck.k9.NotificationVibration
-import com.fsck.k9.VibratePattern
 import org.junit.Test
 import org.mockito.kotlin.mock
