+++ legacy/notification/src/main/java/app/k9mail/legacy/notification/NotificationSettings.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9
+package app.k9mail.legacy.notification
 
 /**
  * Describes how a notification should behave.
